+++ packages/imperative/src/plugins/PluginManagementFacility.ts
@@ -261,7 +261,7 @@ export class PluginManagementFacility {
         // Loop through each overrides setting here. Setting is an override that we are modifying while
         // plugin is the pluginName from which to get the setting. This is probably the ugliest piece
         // of code that I have ever written :/
-         for (const [setting, plugin] of Object.entries(AppSettings.instance.getNamespace("overrides"))) {
+        for (const [setting, plugin] of Object.entries(AppSettings.instance.getNamespace("overrides"))) {
 
             const pluginName = this.getPluginName(plugin, setting);
             if (pluginName !== false && pluginName !== ImperativeConfig.instance.hostPackageName) {